A connected, operating adapter shall not be affected by the scan process.
Extended ScanNetworksAsync by either stating to only scan using specific interfaces, or to scan using only interfaces that are not in use (disconnected).

Reason: If there are e.g. two interface adapters, one is connected and one disconnected, and ScanNetworksAsync is called to find a network to connect to using the disconnected adapter,
also the connected adapter will pause operation for about four seconds in order to switch channels for scanning. This interruption can be avoided.